// ubl-model-library/src/ubl/2.3/model/cac/DebitNoteLine.ts
export interface DebitNoteLine {
  /**
   * A container for extensions foreign to the document.
   * Cardinality: 0..1
   */
  UBLExtensions?: Array<UBLExtensions> | undefined

  /**
   * An identifier for this debit note line.
   * Identifier
   * Cardinality: 1
   */
  ID: Array<Identifier>

  /**
   * A universally unique identifier for this debit note line.
   * Identifier
   * Cardinality: 0..1
   */
  UUID?: Array<Identifier> | undefined

  /**
   * Free-form text conveying information that is not contained explicitly in other structures.
   * Text
   * Cardinality: 0..n
   */
  Note?: Array<Text> | undefined

  /**
   * The quantity of Items debited in this debit note line.
   * Quantity
   * Cardinality: 0..1
   */
  DebitedQuantity?: Array<Quantity> | undefined

  /**
   * The total amount for this debit note line, including allowance charges but net of taxes.
   * Amount
   * Cardinality: 1
   */
  LineExtensionAmount: Array<Amount>

  /**
   * The total amount for this debit note line, including all allowances, charges and taxes.
   * Amount
   * Cardinality: 0..1
   */
  TaxInclusiveLineExtensionAmount?: Array<Amount> | undefined

  /**
   * The date of this debit note line, used to indicate the point at which tax becomes applicable.
   * Date
   * Cardinality: 0..1
   */
  TaxPointDate?: Array<Date> | undefined

  /**
   * The buyer's accounting cost centre for this debit note line, expressed as a code.
   * Code
   * Cardinality: 0..1
   */
  AccountingCostCode?: Array<Code> | undefined

  /**
   * The buyer's accounting cost centre for this debit note line, expressed as text.
   * Text
   * Cardinality: 0..1
   */
  AccountingCost?: Array<Text> | undefined

  /**
   * A code signifying the business purpose for this payment.
   * Code
   * Cardinality: 0..1
   */
  PaymentPurposeCode?: Array<Code> | undefined

  /**
   * A reason for the debit.
   * Response
   * Cardinality: 0..n
   */
  DiscrepancyResponse?: Array<Response> | undefined

  /**
   * A reference to a despatch line associated with this debit note line.
   * Line Reference
   * Cardinality: 0..n
   */
  DespatchLineReference?: Array<LineReference> | undefined

  /**
   * A reference to a receipt line associated with this debit note line.
   * Line Reference
   * Cardinality: 0..n
   */
  ReceiptLineReference?: Array<LineReference> | undefined

  /**
   * A reference to a billing document associated with this debit note line.
   * Billing Reference
   * Cardinality: 0..n
   */
  BillingReference?: Array<BillingReference> | undefined

  /**
   * A reference to a document associated with this debit note line.
   * Document Reference
   * Cardinality: 0..n
   */
  DocumentReference?: Array<DocumentReference> | undefined

  /**
   * A reference to pricing and item location information associated with this debit note line.
   * Pricing Reference
   * Cardinality: 0..1
   */
  PricingReference?: Array<PricingReference> | undefined

  /**
   * A delivery associated with this debit note line.
   * Delivery
   * Cardinality: 0..n
   */
  Delivery?: Array<Delivery> | undefined

  /**
   * A total amount of taxes of a particular kind applicable to this debit note line.
   * Tax Total
   * Cardinality: 0..n
   */
  TaxTotal?: Array<TaxTotal> | undefined

  /**
   * An allowance or charge associated with this debit note.
   * Allowance Charge
   * Cardinality: 0..n
   */
  AllowanceCharge?: Array<AllowanceCharge> | undefined

  /**
   * The item associated with this debit note line.
   * Item
   * Cardinality: 0..1
   */
  Item?: Array<Item> | undefined

  /**
   * The price of the item associated with this debit note line.
   * Price
   * Cardinality: 0..1
   * Alternative business terms: Unit Price, Base Price
   */
  Price?: Array<Price> | undefined

  /**
   * A recursive description of a debit note line subsidiary to this debit note line.
   * Debit Note Line
   * Cardinality: 0..n
   */
  SubDebitNoteLine?: Array<DebitNoteLine> | undefined
}